+++ scripts/git-subtree/template/add.ps1
@@ -1,0 +1,5 @@
+git subtree add `
+    -P workspace/plugins/template `
+    https://github.com/Zuoqiu-Yingyi/siyuan-plugin-template.git `
+    main `
+    --squash