+++ scripts/git-subtree/template/pull.ps1
@@ -1,0 +1,4 @@
+git subtree pull `
+    -P workspace/plugins/template `
+    https://github.com/Zuoqiu-Yingyi/siyuan-plugin-template.git `
+    main