+++ scripts/git-subtree/template/push.ps1
@@ -1,0 +1,4 @@
+git subtree push `
+    -P workspace/plugins/template `
+    https://github.com/Zuoqiu-Yingyi/siyuan-plugin-template.git `
+    main